agri-analysis-server/requirements.txt
cffi==1.14.6
click==8.0.1
cryptography==3.4.7
Flask==2.0.1
Flask-Cors==3.0.10
Flask-SQLAlchemy==2.5.1
greenlet==1.1.0
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
pycparser==2.20
PyMySQL==1.0.2
PyYAML==5.4.1
six==1.16.0
SQLAlchemy==1.4.22
Werkzeug==2.0.1
